fix: Manually fix mapping errors between countries
Some countries were spelled differently (as the ISO input list is
randomly in English/Norwegian).

Some territories of the Netherlands didn't have their own separate
code in the ISO list (which for the purpose of this exercise is the
master data), they were mapped to `NL`.

Kosovo only has a "de-facto" country code and is also not in the list.

The SSB list primarily deals with citizenship. Codes for stateless
people are not mapped.